requirements.txt
python-jobspy
pandas
numpy
pymupdf
nltk
scikit-learn
rapidfuzz
pyarrow
supabase
streamlit
